Adds Penalties, JSON storage.
Penalties are added as an additional customization for Nodes. They do
not contribute to the total cost on individual training samples, but
they add to the gradients for each weight.

JSONAble is added as a new optional interface, available for all
optional interfaces defined by badstudent. JSONAble allows simplified
saving and loading by allowing user-defined interfaces to supply
themselves as enough to be written to a file via encoding/json.

Additionally, a bug is removed in *Node.SetDelay(), which caused
successful execution to return nil in place of the original Node.